v3.0.4
v3.0.4
negative multispleef/multisplegg elo
spectating while spectating says already in battle while spectating while
having random power in power spleef training not swap every time another power is selected (also fixes when we rename powers)
multispleef/multisplegg works
leaderboard searching for player not working
multispleef/multisplegg ratings backwards
builder perms gonezo
fire flinger change (25% particles, 50% sound)
checkpoints
parties removed from queue when a player joins a game
teamspleef 3v3 no worky (maybe fixed, maybe broke other teamspleef queues)

v3.0.3
splegg gun stops working if shooting right at start of round
powerspleef ice pillar -> snow pillar
powerspleef ender rift -> back 3sec to 2sec
powerspleef starfall -> stay at initial area
destroying collectibles also removes all collectibles from offline players that have it and replaces it with a crate

+++ Core/src/main/java/com/spleefleague/core/util/packet/BlockPalette.java
@@ -63,7 +63,7 @@ public abstract class BlockPalette {
             ProtocolLongArrayBitReader reader = new ProtocolLongArrayBitReader(data);
             BlockData[] bdata = new BlockData[4096];//Chunk section is 16x16x16
             for (int i = 0; i < bdata.length; i++) {
-                int id = reader.readInt(14);
+                int id = reader.readShort(14);
                 bdata[i] = blockDataFromId(id);
             }
             return bdata;
